[move] ScalikeJDBCConfigurationのパスを変更する
